nam comiit new

+++ src/main/java/com/example/repository/WardRepository.java
@@ -7,6 +7,6 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface WardRepository extends JpaRepository<Ward, Long> {
-    List<Ward> findByProvince_ProvinceId(Long provinceId);
+public interface WardRepository extends JpaRepository<Ward, Integer> {
+    List<Ward> findByProvince_ProvinceId(Integer provinceId);
 }
